Move reconciliation due tracker to dedicated file
